+++ Nickel/Framework/ModEventManager.cs
@@ -12,12 +12,24 @@ internal sealed class ModEventManager
 	public readonly ManagedEvent<LoadStringsForLocaleEventArgs> OnLoadStringsForLocaleEvent;
 	public readonly ManagedEvent<Exception?> OnGameClosingEvent;
 
+	public ManagedEvent<State> OnSaveLoadedEvent
+	{
+		get
+		{
+			if (this.CurrentModLoadPhaseProvider().Phase < ModLoadPhase.AfterGameAssembly)
+				throw new InvalidOperationException($"Cannot access the {nameof(this.OnSaveLoadedEvent)} event before the game assembly is loaded");
+			if (this.OnSaveLoadedEventStorage is ManagedEvent<State> @event)
+				return @event;
+			throw new InvalidOperationException($"The {nameof(this.OnSaveLoadedEvent)} event should be set up by now, but it is not");
+		}
+	}
+
 	public Artifact PrefixArtifact
 	{
 		get
 		{
 			if (this.CurrentModLoadPhaseProvider().Phase < ModLoadPhase.AfterGameAssembly)
-				throw new InvalidOperationException("Cannot access artifact hooks before the game assembly is loaded.");
+				throw new InvalidOperationException("Cannot access artifact hooks before the game assembly is loaded");
 			if (this.PrefixArtifactStorage is Artifact artifact)
 				return artifact;
 
@@ -42,15 +54,19 @@ internal sealed class ModEventManager
 		}
 	}
 
-	private object? PrefixArtifactStorage { get; set; }
-	private object? SuffixArtifactStorage { get; set; }
+	private object? PrefixArtifactStorage;
+	private object? SuffixArtifactStorage;
+	private object? OnSaveLoadedEventStorage;
+	private WeakReference<object>? LastState;
 
-	private Func<ModLoadPhaseState> CurrentModLoadPhaseProvider { get; }
+	private readonly Func<ModLoadPhaseState> CurrentModLoadPhaseProvider;
+	private readonly Func<IModManifest, ILogger> LoggerProvider;
 
 	public ModEventManager(Func<ModLoadPhaseState> currentModLoadPhaseProvider, Func<IModManifest, ILogger> loggerProvider, IModManifest modLoaderModManifest)
 	{
 		this.ModLoaderModManifest = modLoaderModManifest;
 		this.CurrentModLoadPhaseProvider = currentModLoadPhaseProvider;
+		this.LoggerProvider = loggerProvider;
 		this.OnModLoadPhaseFinishedEvent = new((_, mod, exception) =>
 		{
 			var logger = loggerProvider(mod);
@@ -70,18 +86,26 @@ internal sealed class ModEventManager
 		this.OnModLoadPhaseFinishedEvent.Add(this.OnModLoadPhaseFinished, modLoaderModManifest);
 	}
 
-	[EventPriority(double.MinValue)]
+	[EventPriority(double.MaxValue)]
 	private void OnModLoadPhaseFinished(object? _, ModLoadPhase phase)
 	{
 		if (phase != ModLoadPhase.AfterGameAssembly)
 			return;
-		this.SubscribeAfterGameAssembly();
+		this.SetupAfterGameAssembly();
 	}
 
-	private void SubscribeAfterGameAssembly()
+	private void SetupAfterGameAssembly()
 	{
+		this.OnSaveLoadedEventStorage = new ManagedEvent<State>((_, mod, exception) =>
+		{
+			var logger = this.LoggerProvider(mod);
+			logger.LogError("Mod failed in `{Event}`: {Exception}", nameof(this.OnSaveLoadedEvent), exception);
+		});
+		
 		StatePatches.OnEnumerateAllArtifacts += this.OnEnumerateAllArtifacts;
 		ArtifactPatches.OnKey += this.OnArtifactKey;
+		CheevosPatches.OnCheckOnLoad += this.OnCheckOnLoad;
+		GPatches.OnAfterFrame += this.OnAfterFrame;
 	}
 
 	private void OnEnumerateAllArtifacts(object? _, StatePatches.EnumerateAllArtifactsEventArgs e)
@@ -101,6 +125,21 @@ internal sealed class ModEventManager
 	{
 		if (ReferenceEquals(e.Artifact, this.PrefixArtifact) || ReferenceEquals(e.Artifact, this.SuffixArtifact))
 			e.Key = e.Artifact.GetType().Name;
+	}
+
+	private void OnCheckOnLoad(object? sender, State state)
+	{
+		this.LastState = new(state);
+		this.OnSaveLoadedEvent.Raise(null, state);
+	}
+
+	private void OnAfterFrame(object? sender, G g)
+	{
+		if (this.LastState is not null && (!this.LastState.TryGetTarget(out var lastState) || g.state == lastState))
+			return;
+		
+		this.LastState = new(g.state);
+		this.OnSaveLoadedEvent.Raise(null, g.state);
 	}
 
 	private GeneratedHookableSubclass<Artifact> CreateHookableArtifactSubclass()